test: getUserMe 기반 테스트로 업데이트
- 삭제: getLeaveSummary / getProfile / getLeavePolicy 테스트
  (해당 메서드가 repository에서 제거됨)
- 추가: user_me_repository_test — getMe() 통합 테스트
  (user / leavePolicy / leaveBalance 필드 및 파생값, 에러 케이스 포함)
